test: improving coverage of dns-lookup
PR-URL: https://github.com/nodejs/node/pull/10844
